requirements.txt
pytest==7.1.3
pytest-pycodestyle
pandas
jupyter
flask
cerberus
requests
numpy
Pillow
matplotlib
pymodm
dnspython
